lstm cell was transferred from op/layers.py to frontend/common.py
